feat(colfix): A CSS hack to make equal height cols
To make equally high columns next to each other, use

<div class="row-fluid-colfix">
  <!-- spans go here -->
</div>

When using wells inside a row-fluid-colfix, use them in the
div where span is used and wrap row-fluid-colfix in a div
with class row-fluid-colfix-well.

To put a margin between the columns, use the classes
- margin5px
- margin10px
- margin20px
on the row-fluid-colfix div. If you want different margins,
add them to row-fluid-colfix.less.

Known not to work in IE<=8 and Safari on Windows.
Tested on Chrome/Firefox on Windows/Linux and Safari on Mac,
as well as on IE10 on Windows 8.