refactor(e2e-tests): now using protractor
This removes the old end to end tests and replaces them with equivalent
protractor compartible tests.

To install protractor on your system,
$ npm install protractor
$ ./node_modules/.bin/webdriver-manager update

To run the tests, start the webdriver manager
$ ./node_modules/.bin/webdriver-manager start
and keep it running.

Now run the tests from a different terminal, using
$ ./node_modules/.bin/protractor src/DotHiv/WebsiteCharityBundle/Resources/test/config/protractor-e2e.js
from the root project directory.

Protractor can also be run in debugging mode
$ ./node_modules/.bin/protractor debug src/DotHiv/WebsiteCharityBundle/Resources/test/config/protractor-e2e.js
In debugging mode, the javascript statement
  debugger;
causes the test to stop and launches the CLI step-by-step debugger.

Unfourtunately, tests fail from time to time due to a timeout. This
behaviour seems to be unpredictible and is related to the AJAX calls.
If run in debug mode, the tests seems to be less likely to fail randomly.