feat(dhTeaser): directive for teaser texts.
Usage: <dh-teaser prefix="my.prefix"></dh-teaser>

It will display a paragraph with the text of the translation key
my.prefix.text and a link with the label my.prefix.dropdown.

After clicking on the link, the link will be hidden and a paragraph
with the text my.prefix.expand will be shown additionally to the
first paragraph.

Simple unit tests included. Some unit tests still need to be
implemented.